requirements.txt
web3
solders
requests
python-dotenv

